✨ feat: add price_u64 module for nanodollar precision billing (#44)
* ✨ feat: add price_u64 module for nanodollar precision billing

Task 10.1 & 10.2: Create helper module for u64 price conversions
- Add NANO_PER_DOLLAR and RATE_SCALE constants
- Implement dollars_to_nano and nano_to_dollars converters
- Implement rate_to_scaled and scaled_to_rate for exchange rates
- Implement calculate_cost_safe with u128 overflow protection
- Add comprehensive unit tests for all conversion functions
- Export key functions from common crate lib.rs

+++ crates/client/crates/client-shared/src/services/usage_service.rs
@@ -14,7 +14,8 @@ pub struct UsageStats {
 pub struct Recharge {
     pub id: i32,
     pub user_id: String,
-    pub amount: f64,
+    /// Amount in nanodollars (9 decimal precision)
+    pub amount: i64,
     pub description: Option<String>,
     pub created_at: Option<String>,
 }
